Done and tested reclaim protocol

+++ src/client/ClientInterface.java
@@ -7,6 +7,6 @@ public interface ClientInterface extends Remote {
     boolean backup(String path, int replicationDegree) throws RemoteException;
     boolean restore(String path) throws RemoteException;
     boolean delete(String path) throws RemoteException;
-    boolean reclaim(int diskSpace) throws RemoteException;
+    boolean reclaim(double diskSpace) throws RemoteException;
     String state() throws RemoteException;
 }
